[android] Moved all intent processors from activity to specific intent package to encapsulate it
