+++ android/src/com/mapswithme/maps/MwmActivity.java
@@ -1726,7 +1726,7 @@ public class MwmActivity extends BaseMwmFragmentActivity
     @NonNull
     private final String mUrl;
 
-    ImportBookmarkCatalogueTask(@NonNull String url)
+    public ImportBookmarkCatalogueTask(@NonNull String url)
     {
       mUrl = url;
     }
@@ -1752,7 +1752,7 @@ public class MwmActivity extends BaseMwmFragmentActivity
     private static final long serialVersionUID = 1L;
     private final String mUrl;
 
-    OpenUrlTask(String url)
+    public OpenUrlTask(String url)
     {
       Utils.checkNotNull(url);
       mUrl = url;
@@ -2565,12 +2565,12 @@ public class MwmActivity extends BaseMwmFragmentActivity
     }
   }
 
-  static class ShowPointTask implements MapTask
+  public static class ShowPointTask implements MapTask
   {
     private final double mLat;
     private final double mLon;
 
-    ShowPointTask(double lat, double lon)
+    public ShowPointTask(double lat, double lon)
     {
       mLat = lat;
       mLon = lon;
@@ -2585,7 +2585,7 @@ public class MwmActivity extends BaseMwmFragmentActivity
     }
   }
 
-  static class BuildRouteTask implements MapTask
+  public static class BuildRouteTask implements MapTask
   {
     private final double mLatTo;
     private final double mLonTo;
@@ -2606,18 +2606,18 @@ public class MwmActivity extends BaseMwmFragmentActivity
                                        TextUtils.isEmpty(addr) ? "" : addr, "", lat, lon);
     }
 
-    BuildRouteTask(double latTo, double lonTo, @Nullable String router)
+    public BuildRouteTask(double latTo, double lonTo, @Nullable String router)
     {
       this(latTo, lonTo, null, null, null, null, router);
     }
 
-    BuildRouteTask(double latTo, double lonTo, @Nullable String saddr,
+    public BuildRouteTask(double latTo, double lonTo, @Nullable String saddr,
                    @Nullable Double latFrom, @Nullable Double lonFrom, @Nullable String daddr)
     {
       this(latTo, lonTo, saddr, latFrom, lonFrom, daddr, null);
     }
 
-    BuildRouteTask(double latTo, double lonTo, @Nullable String saddr,
+    public BuildRouteTask(double latTo, double lonTo, @Nullable String saddr,
                    @Nullable Double latFrom, @Nullable Double lonFrom, @Nullable String daddr,
                    @Nullable String router)
     {
