+++ android/src/com/mapswithme/maps/intent/IntentProcessor.java
@@ -1,7 +1,9 @@
-package com.mapswithme.maps;
+package com.mapswithme.maps.intent;
 
 import android.content.Intent;
 import android.support.annotation.NonNull;
+
+import com.mapswithme.maps.MwmActivity;
 
 public interface IntentProcessor
 {
